pyproject.toml
[project]
name = "trendx"
version = "0.1.0"
description = "TrendX Demand Scanner — detect unmet demand before anyone else"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39",
    "httpx>=0.27",
    "pytrends>=4.9",
    "beautifulsoup4>=4.12",
    "rapidfuzz>=3.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.22",
]

[project.scripts]
trendx = "trendx.cli:cli"

[tool.setuptools.packages.find]
include = ["trendx*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
